+++ docs/source/changelog.rst
@@ -3,6 +3,42 @@ Changelog
 
 This document displays the differences between each release of curious.
 
+0.2.0
+-----
+
+ - Add user account logging in support.
+
+ - Add ``_friends`` and ``_blocked`` to :class:`State` to represent the friends and blocked users a client has.
+
+ - Add a new type called :class:`RelationshipUser` which represents either a friend or a blocked user.
+
+ - Rearrange channel and guild handling in ``READY`` parsing.
+
+ - Fix ``author`` inside private DMs being wrong sometimes.
+
+ - Allow group DMs to work properly.
+
+ - User cache has been redesigned:
+
+    - Users are now cached indefinitely in ``state._users``.
+
+    - Users are referred to by property on :class:`Member` rather than by storing them.
+      This should reduce some memory usage as duplicate members will no longer store multiple instances of a user.
+
+    - Users are only decached on a guild member remove.
+
+ - :meth:`State.make_user` now takes a `user_klass` param which allows customization of the user class created when
+   caching a user.
+
+ - Users are now updated in ``PRESENCE_UPDATE`` rather than ``GUILD_MEMBER_UPDATE``.
+
+ - ``GUILD_SYNC`` is now supported for user bots.
+
+ - Creating :class:`HTTPClient` with ``bot=False`` will send a user authorization header rather than a bot
+   authorization header.
+
+ - Add :meth:`HTTPClient.get_user_profile` to get a user's profile.
+
 0.1.4
 -----
 
@@ -13,4 +49,4 @@ This document displays the differences between each release of curious.
  - Fix events inside cogs.
 
  - Add new error code mapping to :class:`HTTPException`. This provides clearer display as to what went wrong when
- performing a HTTP method.
+   performing a HTTP method.